backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "greenlet>=3.2.2",
    "passlib>=1.7.4",
    "pusher>=3.3.3",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "sqlalchemy>=2.0.41",
    "uuid>=1.30",
    "uvicorn>=0.34.2",
]
